dena pasatzen du
datu batetik bestera datu guztiak pasatzen ditu. Eta kodigo guztia metodoetan dago sartuta

+++ Java/bidatubase/src/main/java/controller/SQLserver/BezeroaDao.java
@@ -24,7 +24,7 @@ public class BezeroaDao {
     	return entityManager.createQuery("from Bezeroa").getResultList();
     }
     
-    public Bezeroak getById(long id) {
+    public Bezeroak getById(int id) {
         return entityManager.find(Bezeroak.class, id);
     }
